Handle --check and non-empty fuse mouts
With this change we are able to run this role with the --check flag
without errors.

We are now also able to better ignore the error case where mergerfs
is mounted to a fodler where there already are files (i.e. when we
try to re-mount it). This is not a trivial problem, and it will
probably need some reworking in the future.